4080 【基础信息管理】信息安全管理-机构数据授权：勾选一个区域保存，再次打开，所有区域都是被勾选了。---by zdm

+++ ag-ehr-admin/src/main/java/com/yihu/ehr/organization/controller/OrganizationController.java
@@ -4,12 +4,24 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.yihu.ehr.adapter.service.AdapterOrgClient;
 import com.yihu.ehr.adapter.service.PlanClient;
 import com.yihu.ehr.agModel.geogrephy.GeographyModel;
+import com.yihu.ehr.agModel.org.OrgDetailModel;
+import com.yihu.ehr.agModel.org.OrgModel;
+import com.yihu.ehr.apps.service.AppClient;
 import com.yihu.ehr.constants.AgAdminConstants;
 import com.yihu.ehr.constants.ApiVersion;
+import com.yihu.ehr.controller.BaseController;
+import com.yihu.ehr.fileresource.service.FileResourceClient;
 import com.yihu.ehr.geography.service.AddressClient;
 import com.yihu.ehr.model.adaption.MAdapterOrg;
 import com.yihu.ehr.model.adaption.MAdapterPlan;
+import com.yihu.ehr.model.app.MApp;
+import com.yihu.ehr.model.common.ObjectResult;
+import com.yihu.ehr.model.dict.MConventionalDict;
+import com.yihu.ehr.model.geography.MGeography;
+import com.yihu.ehr.model.geography.MGeographyDict;
+import com.yihu.ehr.model.org.MOrganization;
 import com.yihu.ehr.model.profile.MTemplate;
+import com.yihu.ehr.model.security.MKey;
 import com.yihu.ehr.model.user.MUser;
 import com.yihu.ehr.organization.service.OrganizationClient;
 import com.yihu.ehr.redis.client.RedisUpdateClient;
@@ -18,6 +30,7 @@ import com.yihu.ehr.systemdict.service.ConventionalDictEntryClient;
 import com.yihu.ehr.template.service.TemplateClient;
 import com.yihu.ehr.users.service.UserClient;
 import com.yihu.ehr.util.datetime.DateUtil;
+import com.yihu.ehr.util.rest.Envelop;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
@@ -102,7 +115,7 @@ public class OrganizationController extends BaseController {
         try {
             String address = "";
             if(StringUtils.isNotBlank(province)){
-              List<String> addressList = addressClient.search(province, city, district);
+                List<String> addressList = addressClient.search(province, city, district);
                 String[] addrIdsArrays = addressList.toArray(new String[addressList.size()]);
                 address = String.join(",", addrIdsArrays);
             }
@@ -253,7 +266,7 @@ public class OrganizationController extends BaseController {
                 return failed("删除失败!");
             }
             try {
-               fileResourceClient.filesDelete(orgCode);
+                fileResourceClient.filesDelete(orgCode);
             }catch (Exception e){
                 return success("数据删除成功！图片删除失败！");
             }
@@ -365,71 +378,71 @@ public class OrganizationController extends BaseController {
             @RequestParam(value = "inputStream", required = false) String inputStream,
             @ApiParam(name = "imageName", value = "图片全名", defaultValue = "")
             @RequestParam(value = "imageName", required = false) String imageName) {
-      try {
-          String errorMsg ="";
-          //头像上传,接收头像保存的远程路径  path
-          String path = null;
-          if (!StringUtils.isEmpty(inputStream)) {
-              String jsonData = inputStream + "," + imageName;
-              path = orgClient.uploadPicture(jsonData);
-          }
-          String locationId = null;
-          if (!StringUtils.isEmpty(geographyModelJsonData)) {
-              GeographyModel geographyModel = objectMapper.readValue(geographyModelJsonData, GeographyModel.class);
-              if (geographyModel.nullAddress()) {
-                  errorMsg+="机构地址不能为空！";
-              }
-              locationId = addressClient.saveAddress(objectMapper.writeValueAsString(geographyModel));
-              if (StringUtils.isEmpty(locationId)) {
-                  return failed("保存地址失败！");
-              }
-              if(StringUtils.isNotEmpty(errorMsg))
-              {
-                  return failed(errorMsg);
-              }
-          }
-          OrgDetailModel orgDetailModel = null;
-          if (!StringUtils.isEmpty(mOrganizationJsonData)) {
-              orgDetailModel = objectMapper.readValue(mOrganizationJsonData, OrgDetailModel.class);
-          }
-          if (!StringUtils.isEmpty(path)) {
-              orgDetailModel.setImgRemotePath(path);
-              orgDetailModel.setImgLocalPath("");
-          }
-          MOrganization mOrganization = convertToMOrganization(orgDetailModel);
-          if (StringUtils.isEmpty(mOrganization.getOrgCode())) {
-              errorMsg+="机构代码不能为空！";
-          }
-          if (StringUtils.isEmpty(mOrganization.getFullName())) {
-              errorMsg+="机构全名不能为空！";
-          }
-          if (StringUtils.isEmpty(mOrganization.getShortName())) {
-              errorMsg+="机构简称不能为空！";
-          }
-          if (StringUtils.isEmpty(mOrganization.getTel())) {
-              errorMsg+="联系方式不能为空！";
-          }
-          if(StringUtils.isNotEmpty(errorMsg)) {
-              return failed(errorMsg);
-          }
-          mOrganization.setLocation(locationId);
-          String mOrganizationJson = objectMapper.writeValueAsString(mOrganization);
-          MOrganization mOrgNew = orgClient.update(mOrganizationJson);
-          if (mOrgNew == null) {
-             return  failed("更新失败");
-          }
-          //更新机构名称缓存（原则上如果缓存更新失败不能影响实际更新结果）
-          try {
-              redisUpdateClient.updateOrgName(mOrgNew.getOrgCode());
-              redisUpdateClient.updateOrgArea(mOrgNew.getOrgCode());
-          }catch (Exception e) {
-              e.printStackTrace();
-          }
-          return success(convertToOrgDetailModel(mOrgNew));
-      } catch (Exception e) {
-          e.printStackTrace();
-          return failedSystem();
-      }
+        try {
+            String errorMsg ="";
+            //头像上传,接收头像保存的远程路径  path
+            String path = null;
+            if (!StringUtils.isEmpty(inputStream)) {
+                String jsonData = inputStream + "," + imageName;
+                path = orgClient.uploadPicture(jsonData);
+            }
+            String locationId = null;
+            if (!StringUtils.isEmpty(geographyModelJsonData)) {
+                GeographyModel geographyModel = objectMapper.readValue(geographyModelJsonData, GeographyModel.class);
+                if (geographyModel.nullAddress()) {
+                    errorMsg+="机构地址不能为空！";
+                }
+                locationId = addressClient.saveAddress(objectMapper.writeValueAsString(geographyModel));
+                if (StringUtils.isEmpty(locationId)) {
+                    return failed("保存地址失败！");
+                }
+                if(StringUtils.isNotEmpty(errorMsg))
+                {
+                    return failed(errorMsg);
+                }
+            }
+            OrgDetailModel orgDetailModel = null;
+            if (!StringUtils.isEmpty(mOrganizationJsonData)) {
+                orgDetailModel = objectMapper.readValue(mOrganizationJsonData, OrgDetailModel.class);
+            }
+            if (!StringUtils.isEmpty(path)) {
+                orgDetailModel.setImgRemotePath(path);
+                orgDetailModel.setImgLocalPath("");
+            }
+            MOrganization mOrganization = convertToMOrganization(orgDetailModel);
+            if (StringUtils.isEmpty(mOrganization.getOrgCode())) {
+                errorMsg+="机构代码不能为空！";
+            }
+            if (StringUtils.isEmpty(mOrganization.getFullName())) {
+                errorMsg+="机构全名不能为空！";
+            }
+            if (StringUtils.isEmpty(mOrganization.getShortName())) {
+                errorMsg+="机构简称不能为空！";
+            }
+            if (StringUtils.isEmpty(mOrganization.getTel())) {
+                errorMsg+="联系方式不能为空！";
+            }
+            if(StringUtils.isNotEmpty(errorMsg)) {
+                return failed(errorMsg);
+            }
+            mOrganization.setLocation(locationId);
+            String mOrganizationJson = objectMapper.writeValueAsString(mOrganization);
+            MOrganization mOrgNew = orgClient.update(mOrganizationJson);
+            if (mOrgNew == null) {
+                return  failed("更新失败");
+            }
+            //更新机构名称缓存（原则上如果缓存更新失败不能影响实际更新结果）
+            try {
+                redisUpdateClient.updateOrgName(mOrgNew.getOrgCode());
+                redisUpdateClient.updateOrgArea(mOrgNew.getOrgCode());
+            }catch (Exception e) {
+                e.printStackTrace();
+            }
+            return success(convertToOrgDetailModel(mOrgNew));
+        } catch (Exception e) {
+            e.printStackTrace();
+            return failedSystem();
+        }
     }
 
     /**
@@ -528,8 +541,8 @@ public class OrganizationController extends BaseController {
 
         //获取机构类别字典值
         if(StringUtils.isNotBlank(mOrg.getOrgType())){
-          MConventionalDict orgTypeDict = conDictEntryClient.getOrgType(mOrg.getOrgType());
-          org.setOrgTypeName(orgTypeDict == null ? "" : orgTypeDict.getValue());
+            MConventionalDict orgTypeDict = conDictEntryClient.getOrgType(mOrg.getOrgType());
+            org.setOrgTypeName(orgTypeDict == null ? "" : orgTypeDict.getValue());
         }
         //获取接入方式字典字典值
         if(StringUtils.isNotBlank(mOrg.getSettledWay())){
@@ -587,7 +600,7 @@ public class OrganizationController extends BaseController {
         String[] fields = {"name","pid"};
         String[] values = {name,String.valueOf(code)};
         List<MGeographyDict> geographyDictList = (List<MGeographyDict>) addressClient.getAddressDict(fields,values);
-        return (int)geographyDictList.get(0).getId();
+        return geographyDictList.get(0).getId();
     }
 
     /**
@@ -677,7 +690,7 @@ public class OrganizationController extends BaseController {
         try {
             Map<String, String> key = orgClient.distributeKey(orgCode);
             if (key.size() == 0) {
-               return failed("机构秘钥分发失败!");
+                return failed("机构秘钥分发失败!");
             }
             return success(key);
         }
